refactor: merge startup types into ConversationStartup
Colocate ConversationStartupState, ConversationStartupMetrics, and ConversationStartResult in one file.
