--- Sources/ElevenLabs/Public/Conversation/Startup/ConversationStartResult.swift
@@ -1,4 +1,0 @@
-public struct ConversationStartResult: Equatable, Sendable {
-    public let callInfo: CallInfo
-    public let metrics: ConversationStartupMetrics
-}
